Moved is_not_scalar
