+++ include/boost/simd/arch/common/generic/function/log2.hpp
@@ -15,7 +15,7 @@
 #include <boost/simd/arch/common/detail/scalar/logarithm.hpp>
 #include <boost/simd/function/frexp.hpp>
 #include <boost/simd/function/ilog2.hpp>
-#include <boost/simd/sdk/is_not_scalar.hpp>
+#include <boost/simd/meta/is_not_scalar.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/config.hpp>
